Add flax Linear Module
