+++ e3nn_jax/flax.py
@@ -1,0 +1,3 @@
+from e3nn_jax._src.linear_flax import Linear
+
+__all__ = ["Linear"]
