+++ e3nn_jax/haiku.py
@@ -1,0 +1,3 @@
+from e3nn_jax._src.linear_haiku import Linear
+
+__all__ = ["Linear"]
